LayoutTests/imported/w3c/web-platform-tests/navigation-api/precommit-handler/precommitHandler-redirect-options-expected.txt


FAIL precommitHandler redirect() with options dictionary assert_equals: expected "#redirect2" but got "#push"
